+++ tools/dashboard/functions/payload-github-status.ts
@@ -1,15 +1,6 @@
-import {https, config} from 'firebase-functions';
-import {verify} from 'jsonwebtoken';
-import {setGithubStatus} from './github-status';
-
-/** The repo slug. This is used to validate the JWT is sent from correct repo. */
-const repoSlug = config().repoSlug;
-
-/** API token for the Github repository. Required to set the github status on commits and PRs. */
-const repoToken = config().repoToken;
-
-/** The JWT secret. This is used to validate JWT. */
-const jwtSecret = config().jwtSecret;
+import {https} from 'firebase-functions';
+import {verifyToken} from './jwt/verify-token';
+import {setGithubStatus} from './github/github-status';
 
 export const payloadGithubStatus = https.onRequest(async (request, response) => {
   const authToken = request.header('auth-token');
@@ -28,7 +19,7 @@ export const payloadGithubStatus = https.onRequest(async (request, response) => 
     return response.status(400).json({message: 'No valid payload diff has been specified.'});
   }
 
-  await setGithubStatus(commitSha, repoToken, {
+  await setGithubStatus(commitSha, {
     result: true,
     name: 'Library Payload',
     url: `https://travis-ci.org/angular/material2/jobs/${process.env['TRAVIS_JOB_ID']}`,
@@ -37,15 +28,3 @@ export const payloadGithubStatus = https.onRequest(async (request, response) => 
 
   response.json({message: 'Payload Github status successfully set.'});
 });
-
-function verifyToken(token: string): boolean {
-  try {
-    const tokenPayload = verify(token, jwtSecret, {issuer: 'Travis CI, GmbH'});
-    if (tokenPayload.slug !== repoSlug) {
-      console.log(`JWT slugs are not matching. Expected ${repoSlug}`);
-    }
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
